move plugin utils function to a separate lua file
